requirements.txt
flask==2.3.5
python-dotenv==1.0.0
google-genai==0.6.0  # or google-cloud-aiplatform if google-genai not found
